Tempus: Separate Stepper Header and Source
Cleanup the Stepper base class and separate the header from
the source code.  All tests pass on a Mac.

 #4843
